Move constants to the top level dir, add SD_MOUNT_NAME
