+++ prusa/link/printer_adapter/informers/filesystem/mounts.py
@@ -7,7 +7,7 @@ from blinker import Signal
 
 from prusa.link.config import Config
 from prusa.link.printer_adapter.default_settings import get_settings
-from prusa.link.printer_adapter.structures.constants import BLACKLISTED_PATHS, \
+from prusa.link.printer_adapter.const import BLACKLISTED_PATHS, \
     BLACKLISTED_NAMES, BLACKLISTED_TYPES, QUIT_INTERVAL, DIR_RESCAN_INTERVAL
 from prusa.link.printer_adapter.updatable import ThreadedUpdatable
 from prusa.link.printer_adapter.util import get_clean_path, ensure_directory
