+++ prusa/link/printer_adapter/model.py
@@ -2,7 +2,7 @@ from threading import Lock
 
 
 from prusa.connect.printer.const import State
-from prusa.link.printer_adapter.structures.constants import PRINTING_STATES
+from prusa.link.printer_adapter.const import PRINTING_STATES
 from prusa.link.printer_adapter.structures.mc_singleton import MCSingleton
 from prusa.link.printer_adapter.structures.model_classes import Telemetry
 from prusa.link.printer_adapter.structures.module_data_classes import \
