+++ prusa/link/printer_adapter/print_stats.py
@@ -1,7 +1,7 @@
 import logging
 from time import time
 
-from prusa.link.printer_adapter.structures.constants import TAIL_COMMANDS
+from prusa.link.printer_adapter.const import TAIL_COMMANDS
 from prusa.link.printer_adapter.util import get_gcode
 
 log = logging.getLogger(__name__)
