+++ prusa/link/printer_adapter/updatable.py
@@ -1,7 +1,7 @@
 """Definition for ThreadUpdatable class."""
 from threading import Thread
 
-from prusa.link.printer_adapter.structures.constants import QUIT_INTERVAL
+from prusa.link.printer_adapter.const import QUIT_INTERVAL
 from prusa.link.printer_adapter.util import run_slowly_die_fast
